+++ src/main/java/com/caorderapi/repository/ProductRepository.java
@@ -12,5 +12,5 @@ public interface ProductRepository extends JpaRepository<ProductEntity, UUID> {
 
     @Modifying(clearAutomatically = true, flushAutomatically = true)
     @Query("update ProductEntity p set p.stockQuantity = p.stockQuantity - :qty where p.id = :id and p.stockQuantity >= :qty")
-    int decrementStockIfAvailable(@Param("id") UUID productId, @Param("qty") int quantity);
+    int decrementStockIfAvailable(@Param("id") UUID productId, @Param("qty") long quantity);
 }
